Added main guard
In this commit I have added a main guard to the program and made it so the light should turn off after the data has been sent. I have also added another exception to help catch errors when trying to disconnect and reconnect fast and commented out the data that can be displayed in the terminal.